+++ app/comment/[comment_id]/report/loading.tsx
@@ -1,6 +1,6 @@
-import { VStack, HStack } from '@/components/basic/stack';
-import Tag from '@/components/basic/tag';
-import Skeleton from '@/components/composite/skeleton';
+import { VStack, HStack } from '@/component/basic/stack';
+import Tag from '@/component/basic/tag';
+import Skeleton from '@/component/composite/skeleton';
 
 export default function CommentReportLoading() {
   return (
