+++ app/lecture/[id]/write/page.tsx
@@ -5,11 +5,11 @@ import Link from 'next/link';
 
 import { apiInsertComment, apiStartNewComment } from '@/lib/api/calls/course';
 
-import CommentEditor from '@/components/composite/commentEditor';
-import Button from '@/components/basic/button';
+import CommentEditor from '@/component/composite/commentEditor';
+import Button from '@/component/basic/button';
+import { FinishIcon } from '@/component/icon';
 
 import ErrorTemplate from '@/lib/template';
-import { FinishIcon } from '@/lib/icons';
 import { URL_CUSTOMER_SURVEY } from '@/lib/directive';
 
 function NewCommentWithId(id: number) {
