+++ app/mypage/loading.tsx
@@ -1,6 +1,6 @@
 import { MyProfileBasicsLoading } from './static/MyProfileBasics';
 import { MyProfileMembershipsLoading } from './static/MyProfileMemberships';
-import Spinner from '@/components/basic/spinner';
+import Spinner from '@/component/basic/spinner';
 
 export default function MyPageLoading() {
   return (
